refactor: extend k8sfilterhighlight from common
