+++ src/main/java/com/redhat/devtools/intellij/tektoncd/ui/editors/TektonFileEditorProvider.java
@@ -22,7 +22,7 @@ import com.intellij.openapi.util.Key;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiFile;
 import com.intellij.psi.PsiManager;
-import com.redhat.devtools.intellij.tektoncd.validation.KubernetesTypeInfo;
+import com.redhat.devtools.intellij.common.validation.KubernetesTypeInfo;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
